refactor: ColorTags 디렉토리 위치 수정 및 css styled 내부 처리
